changed divide return value to double

+++ src/main/java/MyMathCalc.java
@@ -8,7 +8,7 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public float divide(float a, float b) {
+    public double divide(double a, double b) {
         return a / b;
     }
 }
